refactor: fga-eps-mds/2020.2-Lend.it#110 Update widgets names and fix imports
